# src/backend/cube/docs/swagger.yaml
basePath: /api/v1
definitions:
  AdminBlockUserResponse:
    properties:
      blocked_until:
        example: "2025-08-09T15:00:00.053Z"
        type: string
      id:
        example: f6457bdf-4e67-4f05-9108-1cbc0fec9405
        type: string
    type: object
  AdminModelInfoResponse:
    properties:
      id:
        example: f6457bdf-4e67-4f05-9108-1cbc0fec9405
        type: string
      name:
        example: awesome_username
        type: string
      structure: {}
    type: object
  AdminModelWeightsInfoResponse:
    properties:
      id:
        example: f6457bdf-4e67-4f05-9108-1cbc0fec9405
        type: string
      name:
        example: awesome_username
        type: string
      offsets: {}
      weights: {}
    type: object
  AdminUserInfoResponse:
    properties:
      email:
        example: my_awesome@email.com
        type: string
      fullname:
        example: Ivanov Ivan Ivanovich
        type: string
      id:
        example: f6457bdf-4e67-4f05-9108-1cbc0fec9405
        type: string
      username:
        example: awesome_username
        type: string
    type: object
  LoginRequest:
    properties:
      email:
        example: my_awesome@email.com
        type: string
      password:
        example: Really, you're waiting for example?
        type: string
    required:
    - email
    - password
    type: object
  LoginResponse:
    properties:
      expire:
        example: "2022-03-20T17:00:01Z"
        type: string
      token:
        example: eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2NDc...
        type: string
    type: object
...
  Unauthorized:
    properties:
      message:
        example: user not found
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is cube server.
  license:
    name: MIT
    url: https://mit-license.org/
  title: Cube API
  version: "1.0"
paths:
  /api/v1/admin/login:
    post:
      consumes:
      - application/json
      description: login to existing account
      parameters:
      - description: The body to create a thing
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      responses:
        "200":
          description: Login was successfull
          schema:
            $ref: '#/definitions/LoginResponse'
        "401":
          description: Login data is invalid or missing, check request
          schema:
            $ref: '#/definitions/Unauthorized'
      summary: Admin login
      tags:
      - auth
      - admin
  /api/v1/admin/models:
    delete:
      description: Deletes model info from any user
      parameters:
      - description: Model ID to delete
        in: query
        name: id
        type: string
      responses:
        "200":
          description: Model info deleted
        "400":
          description: Invalid request
        "500":
          description: Failed to delete model info from storage
      summary: Delete model info
      tags:
      - admin
    get:
      description: Find such model info as id, username, email and fullname
      parameters:
      - description: Model ID to search for
        in: query
        name: id
        type: string
      - description: User ID that owns model to search for
        in: query
        name: owner_id
        type: string
      - description: Model name to search for
        in: query
        name: name
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Page size for pagination
        in: query
        name: per_page
        type: integer
      responses:
        "200":
          description: Model info found
          schema:
            items:
              $ref: '#/definitions/AdminModelInfoResponse'
            type: array
        "400":
          description: Invalid request
        "500":
          description: Failed to get model info from storage
      summary: Find model info
      tags:
      - admin
  /api/v1/admin/models/weights:
    delete:
      description: Deletes model info from any user
      parameters:
      - description: Model ID to delete
        in: query
        name: id
        type: string
      responses:
        "200":
          description: Model info deleted
        "400":
          description: Invalid request
        "403":
          description: Invalid token, user id not specified
        "500":
          description: Failed to delete model info from storage
      summary: Delete model info
      tags:
      - admin
    get:
      description: Find such model info as id, username, email and fullname
      parameters:
      - description: Weight ID to search for
        in: query
        name: id
        type: string
      - description: Structure ID to search for
        in: query
        name: structure_id
        type: string
      - description: Weights name to search for
        in: query
        name: name
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Page size for pagination
        in: query
        name: per_page
        type: integer
      responses:
        "200":
          description: Model weights info found
          schema:
            items:
              $ref: '#/definitions/AdminModelWeightsInfoResponse'
            type: array
        "400":
          description: Invalid request
        "500":
          description: Failed to get model weights info from storage
      summary: Find model info
      tags:
      - admin
  /api/v1/admin/users:
    delete:
      consumes:
      - application/json
      description: Deletes user by id
      parameters:
      - description: UserId to delete
        in: query
        name: id
        type: string
      responses:
        "200":
          description: User deleted
        "400":
          description: Invalid request
        "500":
          description: Failed to delete user info from storage
      summary: Delete user info
      tags:
      - admin
    get:
      consumes:
      - application/json
      description: Find such users info as id, username, email and fullname
      parameters:
      - description: UserId to search for
        in: query
        name: user_id
        type: string
      - description: Username to search for
        in: query
        name: username
        type: string
      - description: Email to search for
        in: query
        name: email
        type: string
      - description: Email to search for
        in: query
        name: page
        type: integer
      - description: Email to search for
        in: query
        name: per_page
        type: integer
      responses:
        "200":
          description: Users info found
          schema:
            items:
              $ref: '#/definitions/AdminUserInfoResponse'
            type: array
        "400":
          description: Invalid request
        "500":
          description: Failed to get user info from storage
      summary: Find user info
      tags:
      - admin
  /api/v1/admin/users/blocked:
    delete:
      description: Deletes user block info by user id
      parameters:
      - description: User ID to unblock
        in: query
        name: id
        type: string
      responses:
        "200":
          description: User unblocked
        "400":
          description: Invalid request
        "500":
          description: Failed to delete user block info from storage
      summary: Delete user block info
      tags:
      - admin
    get:
      description: Find such users info as id and block time
      parameters:
      - description: UserId to search for
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Users block info found
          schema:
            $ref: '#/definitions/AdminBlockUserResponse'
        "400":
          description: Invalid request
        "500":
          description: Failed to get user info from storage
      summary: Find user block info
      tags:
      - admin
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Login to existing account
      parameters:
      - description: The body to create a thing
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      responses:
        "200":
          description: Login was successfull
          schema:
            $ref: '#/definitions/LoginResponse'
        "401":
          description: Login data is invalid or missing, check request
          schema:
            $ref: '#/definitions/Unauthorized'
      summary: User login
      tags:
      - auth
      - user
...
  /api/v1/stat/login:
    post:
      consumes:
      - application/json
      description: login to existing account
      parameters:
      - description: The body to create a thing
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      responses:
        "200":
          description: Login was successfull
          schema:
            $ref: '#/definitions/LoginResponse'
        "401":
          description: Login data is invalid or missing, check request
          schema:
            $ref: '#/definitions/Unauthorized'
      summary: Stat login
      tags:
      - auth
      - stat
...
securityDefinitions:
  Bearer:
    type: basic
swagger: "2.0"